requirements.txt
python-telegram-bot
deepgram-sdk
stability-sdk